refactor so we can access parser
To inspect the argument parse we need to be able to import it.
